requirements.txt
tqdm==4.63.1
tensorflow==2.8.0
torch==1.6.0
scikit_image==0.19.2
albumentations==1.1.0
scipy==1.8.0
imageio==2.16.1
numpy==1.21.2
torchvision==0.14.1
matplotlib==3.5.1
h5py==3.6.0
opencv_python==4.5.5.64
ipython==8.5.0
Pillow==9.1.0
scikit_learn==0.23.2
scikit-image==0.19.2
tensorboardX==2.5.1